requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
numpy
python-dateutil
python-dotenv
asyncio-mqtt
pyserial
